move to settings.js
