+++ main/index.js
@@ -29,6 +29,7 @@ const { ipcMain } = require('electron/main')
 const os = require('os')
 const core = require('./core')
 const wallet = require('./wallet')
+const settings = require('./settings')
 const serve = require('electron-serve')
 const { setupAppMenu } = require('./app-menu')
 const setupTray = require('./tray')
@@ -176,6 +177,7 @@ async function run () {
     await wallet.setup(ctx)
     await telemetry.setup()
     await core.setup(ctx)
+    await settings.setup(ctx)
   } catch (e) {
     handleError(e)
   }
